Phase 4: termination conditions, flavor text and cleanup
